requirements-ml.txt
pandas
numpy
scikit-learn
joblib
psycopg[binary]
python-dotenv
fairlearn            # opcional (métricas de equidad)
